backend(assets): rename functions
